django_api/requirements.txt
asgiref==3.6.0
Django==4.1.2
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
gunicorn==20.1.0
pymongo==3.11.4
pytz==2023.3
sqlparse==0.2.4